oop hw1 completed

+++ src/com/company/start/hw2/Circle.java
@@ -5,14 +5,14 @@ import java.util.Scanner;
 public class Circle {
     public static void main(String[] args) {
         double radius = 4d;
-        int x;
-        int y;
+        double x;
+        double y;
 
         Scanner scanner = new Scanner(System.in);
         System.out.println("Input x");
-        x = scanner.nextInt();
+        x = scanner.nextDouble();
         System.out.println("Input y");
-        y = scanner.nextInt();
+        y = scanner.nextDouble();
 
         if (Math.sqrt(x * x + y * y) <= radius) {
             System.out.println("Point with coord x = " + x + ", y = " + y
